+++ extensions/editor/browser/index.js
@@ -9,6 +9,9 @@
 import {LAYER, SYSTEM_FIELD_PREFIX} from '../../../browser/modules/layerTree/constants';
 import dayjs from 'dayjs';
 
+const jquery = require('jquery');
+require('snackbarjs');
+
 /**
  *
  * @type {*|exports|module.exports}
@@ -50,6 +53,7 @@ const widgets = {'imageupload': ImageUploadWidget};
 const MODULE_NAME = `editor`;
 const EDITOR_FORM_CONTAINER_ID = 'editor-attr-form';
 const EDITOR_CONTAINER_ID = 'editor-attr-dialog';
+const MAX_NODE_IN_FEATURE = 2000; // If number of nodes exceed this number, when the geometry editor is not enabled.
 
 const serviceWorkerCheck = () => {
     if (!('serviceWorker' in navigator) || !navigator.serviceWorker || !navigator.serviceWorker.controller) {
@@ -699,7 +703,21 @@ module.exports = {
                     break;
 
                 default:
-                    editor = e.enableEdit();
+
+                    let numberOfNodes = 0;
+                    editedFeature.feature.geometry.coordinates.forEach((c) => {
+                        numberOfNodes += c.length;
+                    })
+                    if (numberOfNodes <= MAX_NODE_IN_FEATURE) {
+                        editor = e.enableEdit();
+                    } else {
+                        editor = false;
+                        jquery.snackbar({
+                            content: `<span>${__("Editing of geometry is not possible when number of nodes exceed")} ${MAX_NODE_IN_FEATURE}</span>`,
+                            htmlAllowed: true,
+                            timeout: 6000
+                        })
+                    }
                     break;
             }